requirements.txt
pyspark
requests
pytest
flake8
pylint
pyyaml